requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
alembic==1.10.3
passlib[bcrypt]==1.7.4
